requirements.txt
python-dotenv
requests
notion-client
pyqt5
qtawesome